lab_08/pyproject.toml
[tool.poetry]
name = "lab-08"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.0"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
matplotlib = "^3.9.4"
scikit-image = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
